WLT-471: refactored wallet; returned util from rd; added loop to sh
