+++ ci/local-dev/manifests/ks/components/insolar/launch.sh
@@ -16,16 +16,10 @@ then
     echo "generate members keys in dir: $CONFIG_DIR"
     insolar gen-key-pair > ${CONFIG_DIR}/root_member_keys.json
     insolar gen-key-pair > ${CONFIG_DIR}/migration_admin_member_keys.json
-    insolar gen-key-pair > ${CONFIG_DIR}/migration_daemon_0_member_keys.json
-    insolar gen-key-pair > ${CONFIG_DIR}/migration_daemon_1_member_keys.json
-    insolar gen-key-pair > ${CONFIG_DIR}/migration_daemon_2_member_keys.json
-    insolar gen-key-pair > ${CONFIG_DIR}/migration_daemon_3_member_keys.json
-    insolar gen-key-pair > ${CONFIG_DIR}/migration_daemon_4_member_keys.json
-    insolar gen-key-pair > ${CONFIG_DIR}/migration_daemon_5_member_keys.json
-    insolar gen-key-pair > ${CONFIG_DIR}/migration_daemon_6_member_keys.json
-    insolar gen-key-pair > ${CONFIG_DIR}/migration_daemon_7_member_keys.json
-    insolar gen-key-pair > ${CONFIG_DIR}/migration_daemon_8_member_keys.json
-    insolar gen-key-pair > ${CONFIG_DIR}/migration_daemon_9_member_keys.json
+    for (( b = 0; b < 10; b++ ))
+    do
+    insolar gen-key-pair > ${CONFIG_DIR}/migration_daemon_${b}_member_keys.json
+    done
 
     echo "generate bootstrap files"
     mkdir -vp $CERTS_KEYS
